Move Language.Haskell.LSP.Types to haskell-lsp-types
