+++ haskell-lsp-types/src/Language/Haskell/LSP/Types.hs
@@ -1,0 +1,8 @@
+module Language.Haskell.LSP.Types
+  (
+    module Language.Haskell.LSP.Types.DataTypesJSON
+  , module Language.Haskell.LSP.Types.Lens
+  ) where
+
+import Language.Haskell.LSP.Types.DataTypesJSON
+import Language.Haskell.LSP.Types.Lens
