+++ haskell-lsp-types/src/Language/Haskell/LSP/Types/CodeAction.hs
@@ -1,21 +1,21 @@
 {-# LANGUAGE DuplicateRecordFields      #-}
 {-# LANGUAGE OverloadedStrings          #-}
 {-# LANGUAGE TemplateHaskell            #-}
-module Language.Haskell.LSP.TH.CodeAction where
+module Language.Haskell.LSP.Types.CodeAction where
 
 import           Control.Applicative
 import qualified Data.Aeson                    as A
 import           Data.Aeson.TH
 import           Data.Aeson.Types
 import           Data.Text                      ( Text )
-import           Language.Haskell.LSP.TH.Command
-import           Language.Haskell.LSP.TH.Constants
-import           Language.Haskell.LSP.TH.Diagnostic
-import           Language.Haskell.LSP.TH.List
-import           Language.Haskell.LSP.TH.Location
-import           Language.Haskell.LSP.TH.Message
-import           Language.Haskell.LSP.TH.TextDocument
-import           Language.Haskell.LSP.TH.WorkspaceEdit
+import           Language.Haskell.LSP.Types.Command
+import           Language.Haskell.LSP.Types.Constants
+import           Language.Haskell.LSP.Types.Diagnostic
+import           Language.Haskell.LSP.Types.List
+import           Language.Haskell.LSP.Types.Location
+import           Language.Haskell.LSP.Types.Message
+import           Language.Haskell.LSP.Types.TextDocument
+import           Language.Haskell.LSP.Types.WorkspaceEdit
 
 
 {-
