+++ haskell-lsp-types/src/Language/Haskell/LSP/Types/DocumentFilter.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE TemplateHaskell            #-}
-module Language.Haskell.LSP.TH.DocumentFilter where
+module Language.Haskell.LSP.Types.DocumentFilter where
 
 import           Data.Aeson.TH
 import           Data.Text                      ( Text )
-import           Language.Haskell.LSP.TH.Constants
-import           Language.Haskell.LSP.TH.List
+import           Language.Haskell.LSP.Types.Constants
+import           Language.Haskell.LSP.Types.List
 
 -- ---------------------------------------------------------------------
 {-
