+++ src/Language/Haskell/LSP/Messages.hs
@@ -2,14 +2,14 @@
 {-# LANGUAGE DeriveGeneric #-}
 
 module Language.Haskell.LSP.Messages
-  ( module Language.Haskell.LSP.TH.MessageFuncs
+  ( module Language.Haskell.LSP.Types.MessageFuncs
   , FromClientMessage(..)
   , FromServerMessage(..)
   )
 where
 
-import           Language.Haskell.LSP.TH.MessageFuncs
-import           Language.Haskell.LSP.TH.DataTypesJSON
+import           Language.Haskell.LSP.Types.MessageFuncs
+import           Language.Haskell.LSP.Types
 import           GHC.Generics
 import           Data.Aeson
