--- src/Language/Haskell/LSP/Types.hs
@@ -1,8 +1,0 @@
-module Language.Haskell.LSP.Types
-  (
-    module Language.Haskell.LSP.TH.DataTypesJSON
-  , module Language.Haskell.LSP.TH.Lens
-  ) where
-
-import Language.Haskell.LSP.TH.DataTypesJSON
-import Language.Haskell.LSP.TH.Lens
